refactor(#zimic): improve server directory structure
