+++ packages/zimic/src/interceptor/http/interceptorWorker/__tests__/HttpinterceptorWorker.node.test.ts
@@ -1,6 +1,6 @@
 import { describe } from 'vitest';
 
-import Server from '@/cli/server/Server';
+import Server from '@/server/Server';
 import { getNodeAccessResources } from '@tests/utils/workers';
 
 import { declareSharedHttpInterceptorWorkerTests } from './shared/workerTests';
