+++ packages/zimic/tests/globalSetup/serverOnBrowser.ts
@@ -1,4 +1,4 @@
-import type Server from '@/cli/server/Server';
+import type Server from '@/server/Server';
 
 let server: Server | undefined;
 
@@ -6,7 +6,7 @@ export const GLOBAL_SETUP_SERVER_HOSTNAME = 'localhost';
 export const GLOBAL_SETUP_SERVER_PORT = 3001;
 
 export async function setup() {
-  const Server = (await import('@/cli/server/Server')).default;
+  const Server = (await import('@/server/Server')).default;
 
   server = new Server({
     hostname: GLOBAL_SETUP_SERVER_HOSTNAME,
